+++ protos/data-node/api/v2/trading_data_grpc.pb.go
@@ -142,6 +142,10 @@ type TradingDataServiceClient interface {
 	// - "network"
 	// - "0000000000000000000000000000000000000000000000000000000000000000", the public key for the global rewards account
 	ListTransfers(ctx context.Context, in *ListTransfersRequest, opts ...grpc.CallOption) (*ListTransfersResponse, error)
+	// Get transfer by ID
+	//
+	// Get a specific transfer by ID
+	GetTransferByID(ctx context.Context, in *GetTransferByIDRequest, opts ...grpc.CallOption) (*GetTransferByIDResponse, error)
 	// Get network limits
 	//
 	// Get the network limits relating to asset and market creation
@@ -899,6 +903,15 @@ func (c *tradingDataServiceClient) GetMarketDataHistoryByID(ctx context.Context,
 func (c *tradingDataServiceClient) ListTransfers(ctx context.Context, in *ListTransfersRequest, opts ...grpc.CallOption) (*ListTransfersResponse, error) {
 	out := new(ListTransfersResponse)
 	err := c.cc.Invoke(ctx, "/datanode.api.v2.TradingDataService/ListTransfers", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *tradingDataServiceClient) GetTransferByID(ctx context.Context, in *GetTransferByIDRequest, opts ...grpc.CallOption) (*GetTransferByIDResponse, error) {
+	out := new(GetTransferByIDResponse)
+	err := c.cc.Invoke(ctx, "/datanode.api.v2.TradingDataService/GetTransferByID", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -1945,6 +1958,10 @@ type TradingDataServiceServer interface {
 	// - "network"
 	// - "0000000000000000000000000000000000000000000000000000000000000000", the public key for the global rewards account
 	ListTransfers(context.Context, *ListTransfersRequest) (*ListTransfersResponse, error)
+	// Get transfer by ID
+	//
+	// Get a specific transfer by ID
+	GetTransferByID(context.Context, *GetTransferByIDRequest) (*GetTransferByIDResponse, error)
 	// Get network limits
 	//
 	// Get the network limits relating to asset and market creation
@@ -2404,6 +2421,9 @@ func (UnimplementedTradingDataServiceServer) GetMarketDataHistoryByID(context.Co
 }
 func (UnimplementedTradingDataServiceServer) ListTransfers(context.Context, *ListTransfersRequest) (*ListTransfersResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListTransfers not implemented")
+}
+func (UnimplementedTradingDataServiceServer) GetTransferByID(context.Context, *GetTransferByIDRequest) (*GetTransferByIDResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetTransferByID not implemented")
 }
 func (UnimplementedTradingDataServiceServer) GetNetworkLimits(context.Context, *GetNetworkLimitsRequest) (*GetNetworkLimitsResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetNetworkLimits not implemented")
@@ -3086,6 +3106,24 @@ func _TradingDataService_ListTransfers_Handler(srv interface{}, ctx context.Cont
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(TradingDataServiceServer).ListTransfers(ctx, req.(*ListTransfersRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _TradingDataService_GetTransferByID_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(GetTransferByIDRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(TradingDataServiceServer).GetTransferByID(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/datanode.api.v2.TradingDataService/GetTransferByID",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(TradingDataServiceServer).GetTransferByID(ctx, req.(*GetTransferByIDRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -4614,6 +4652,10 @@ var TradingDataService_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "ListTransfers",
 			Handler:    _TradingDataService_ListTransfers_Handler,
+		},
+		{
+			MethodName: "GetTransferByID",
+			Handler:    _TradingDataService_GetTransferByID_Handler,
 		},
 		{
 			MethodName: "GetNetworkLimits",